fix l_area_area triggers
when inserting into l_area_area while importing data i got the message that relation "link" is not found.
after figuring out the issue (the triggers do not have fully-qualified name of 'link' table and the search path does not contain the musicbrainz schema), decided to add the musicbrainz schema to the search path to fix it (if musicbrainz schema is not ignored).