Make windows-compatible, fix error on import, fix triggers (#4)
* remove bash from run_script

instead set shell=true. this should not break Linux but should make the script work on Windows, too

* Update replication.py

fix connect_db call

* fix l_area_area triggers

when inserting into l_area_area while importing data i got the message that relation "link" is not found.
after figuring out the issue (the triggers do not have fully-qualified name of 'link' table and the search path does not contain the musicbrainz schema), decided to add the musicbrainz schema to the search path to fix it (if musicbrainz schema is not ignored).

* remove unused param p

added by accident

* Revert "fix l_area_area triggers"

This reverts commit 633fb10e41edbcdd1c4926d856071f63c5642b0b.

* fix l_area_area triggers

now the more "correct" way to use what is already there.