Use v2 packets (#5)
This modifies mbslave to use the v2 packets produced by
https://github.com/metabrainz/dbmirror2.

These have been used by default in MusicBrainz Server's replication system
since the last schema change.

I removed the custom parsing code and just load the pending_* files into the
dbmirror2 schema via COPY. The reason I did this is that the parsing of escape
sequences is more complex now that data is stored in the packets as JSON:
something like "\\n" in the JSON (which represents a single newline) would be
converted to "\<LF>" by the current parsing code, which is an invalid input to
the JSON parser. I tried to fix this, but it's less error-prone to let postgres
do the parsing.